Refactor for clarity and add missing spread
